Implemented (but did not test) the SendableChooser
In this branch, there is no test for the SendableChooser. To test the SendableChooser, one must merge this branch into his branch and create the test in that branch. Consequently, I have commented out the sample lines in Robot.cpp that would have added both a default autonomous mode and a second autonomous mode to the instance of the SendableChooser.