Began adding purpose-describing enumerators to class autoStrafe
These enumerators are meant to describe the purpose of any particular call of the autoStrafe command so that the command can react specifically to the enumerator and supply the appropriate constants (e.g. drive distance, timeout). This will allow preferences to be used for constants such as the drive distance and timeout of any call of autoStrafe. If a similar use of enumerators is applied to the elevator commands, the number of elevator-related commands could be reduced to one, which would help because the number of commands we have seems excessive. This use of enumerators should also be applied to class autoDrive and perhaps even autonomous group commands.