Added the commits from the other unsynced laptop to this current branch/repository
The first area of changes is in Robot.cpp, where I added a printing of the Sendable Chooser in both Robot::AutonomousInit() and Robot::TeleopInit(). This is to make it unnecessary to reboot the roboRIO to make the list of autonomous modes appear.
The second (and greater) area of changes is in autoStrafeRightThreeTote.cpp. Because of the fifteen second time limit, the robot had stopped just before picking up the third tote. Consequently, I had to take out parts of it to save time. So, the consequent changes were: 1) the robot no longer strafes to the right before moving the elevator up after picking up a yellow tote (this doesn't seem necessary, since the robot knocks the container out of the way), 2) the robot should instead push the third yellow tote into the Auto Zone, 3) as the robot does that described in the latter, it should be putting down the other two yellow totes on top of the third one, and 4) the robot should stop on the scoring platform, leaving the stack of three totes just barely in the Auto Zone (for this step, I had to guess the distance).