Changed the distance to drive forward with the three yellow tote stack and made the robot back up at the end
The robot hasn't gotten that far in the autonomous mode. However, in the recent match, the robot failed to pick up the yellow tote, so it kept sensing the yellow tote and didn't strafe. The robot ended up driving forward, and it was there we could see the robot wouldn't have made it into the auto zone anyway.